+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.1
 ---
 
+ * Add `$tagsToKeep` argument to `DecoratorServicePass`
  * Support autowiring env vars as closures or `Stringable` when using `#[Autowire(env: 'FOO')]`
  * Add `EnvClosureArgument` and `!env_closure` YAML tag to inject env vars as closures or `Stringable` arguments
  * Add `AddBehaviorDescribingTagsPass` to allow bundles to extend the list of behavior-describing tags